1) Added code to correct capitalization problems with index column names. If a column name in an index does not exactly match the name of any column in the table, but does match a column name if matching is made case insensitive, the index column name is replaced by the case-sensitive name of the matching table column. This was a problem with DB, but not with MDB2.
2)  Added $col, $idx, $auto_inc_col columns with formats similar to those
of corresponding properties in DB_Table, rather than properties with
the formats that are output by tableInfo and getIndexes. Moved the code
to put the column and index data into this format into the getTableDefinition
method. The buildTableClass now contains only code to format the string
representation of the class definition.

3) Added auto-magic guessing of foreign key references to the
->generateDatabaseFile method, based on a convention for the column names:
If an integer column name ends with 'id' (case insensitive), the rest of
the name matches the name of another table (case insensitive), and it
passes some other sanity checks, the column is assumed to be a foreign
key reference to the primary key of that table.


git-svn-id: http://svn.php.net/repository/pear/packages/DB_Table/trunk@232217 c90b9560-bf6c-de11-be94-00142212c4b1